Move AuthConfigs to default auth plugin impl module.
